# docs/swagger.yaml
definitions:
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
  scheduled_models.Task:
    properties:
      description:
        type: string
      id:
        type: string
      is_enabled:
        type: boolean
      schedule:
        type: string
    type: object
  scheduled_models.TaskLog:
    properties:
      date_end:
        type: string
      date_start:
        type: string
      error_message:
        type: string
      is_good:
        type: boolean
      task_id:
        type: string
    type: object
  status.ApplyBackupRequest:
    properties:
      strategy:
        enum:
        - skip
        - replace
        - fail
        type: string
    required:
    - strategy
    type: object
  status.BackupDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      migration_version:
        type: string
      size:
        type: string
    type: object
  status.Page-status_BackupDTO:
    properties:
      first:
        type: boolean
      items:
        items:
          $ref: '#/definitions/status.BackupDTO'
        type: array
      last:
        type: boolean
      max_page:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      visible:
        type: integer
    type: object
  status.ResponseStatus:
    properties:
      database:
        type: string
      service:
        type: string
    type: object
  status.Version:
    properties:
      applied_at:
        type: string
      id:
        type: string
    type: object
  tasks.UpdateTaskRequest:
    properties:
      is_enabled:
        type: boolean
      schedule:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/status/backups:
    get:
      description: Получить список бэкапов
      parameters:
      - example: '''or'' || ''and'''
        in: query
        name: filters_and_or_switch
        type: string
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        minimum: 1
        name: size
        type: integer
      - example: '''-name,id'' --> name desc and id asc'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Page-status_BackupDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
    post:
      description: Создать бэкап
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.BackupDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
  /api/v1/status/backups/{id}/apply:
    post:
      consumes:
      - application/json
      description: Применить бэкап данных
      parameters:
      - description: ID бэкапа
        in: path
        name: id
        required: true
        type: string
      - description: Стратегия восстановления (skip, replace, fail)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/status.ApplyBackupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
  /api/v1/status/backups/{id}/data:
    get:
      description: Получить сохраненные данные бэкапа
      parameters:
      - description: id бэкапа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
  /api/v1/status/check:
    get:
      description: Статус работоспособности
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.ResponseStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
  /api/v1/status/logs:
    get:
      description: Получить логи
      produces:
      - application/json
      responses:
        "200":
          description: Plain text log file
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
  /api/v1/status/versions:
    get:
      description: Список версий структуры данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/status.Version'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - status
  /api/v1/tasks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scheduled_models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - tasks
  /api/v1/tasks/{id}:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduled_models.Task'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - tasks
    patch:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task update body
        in: body
        name: updateTaskBody
        required: true
        schema:
          $ref: '#/definitions/tasks.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Task ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - tasks
  /api/v1/tasks/{id}/logs:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scheduled_models.TaskLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.MessageResponse'
      tags:
      - tasks
swagger: "2.0"